* include/ruby/encoding.h (rb_enc_step_back): cast 4th argument 'n'   as int because Ruby usually treats length value as long but   onigenc_step_back's 4th argument is int.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@30768 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/encoding.h
@@ -162,7 +162,7 @@ int rb_enc_codelen(int code, rb_encoding *enc);
 /* start, ptr, end, encoding -> next_char */
 #define rb_enc_left_char_head(s,p,e,enc) (char *)onigenc_get_left_adjust_char_head(enc,(UChar*)(s),(UChar*)(p),(UChar*)(e))
 #define rb_enc_right_char_head(s,p,e,enc) (char *)onigenc_get_right_adjust_char_head(enc,(UChar*)(s),(UChar*)(p),(UChar*)(e))
-#define rb_enc_step_back(s,p,e,n,enc) (char *)onigenc_step_back(enc,(UChar*)(s),(UChar*)(p),(UChar*)(e),(n))
+#define rb_enc_step_back(s,p,e,n,enc) (char *)onigenc_step_back(enc,(UChar*)(s),(UChar*)(p),(UChar*)(e),(int)(n))
 
 /* ptr, ptr, encoding -> newline_or_not */
 #define rb_enc_is_newline(p,end,enc)  ONIGENC_IS_MBC_NEWLINE(enc,(UChar*)(p),(UChar*)(end))
